fix: correct EIP-712 signing, regenerate proto types, add missing REST endpoints
Order.marketIndex uint8→uint16 matching Go verifier. Register struct
adds missing sender field (4-field struct). Regenerate proto types
fixing FeeTier tag swap (maker/taker silently swapped) and adding
is_adl/is_force_closed to Order.

Add SetLeverage, SetMarginMode, TransferMargin portfolio endpoints
and GetFeeTiers general endpoint. Add BookView checksum field.

Update staging domain to Monad testnet (chain 10143) and REST URL
to nova.staging.obsdn.trade. Add danger_accept_invalid_certs builder
option for staging TLS.

E2E verified against live staging: register signer, faucet, place
order, cancel order, set leverage.

+++ src/sign/order.rs
@@ -18,7 +18,7 @@ sol! {
     #[derive(Debug)]
     struct Order {
         address sender;
-        uint8 marketIndex;
+        uint16 marketIndex;
         uint8 side;
         uint128 size;
         uint128 price;
@@ -52,7 +52,7 @@ pub struct OrderPayload {
     /// Sender address (the trader's main wallet).
     pub sender: Address,
     /// Market index — looked up from `markets()`/proto `Market.idx`.
-    pub market_index: u8,
+    pub market_index: u16,
     /// Buy / sell.
     pub side: OrderSide,
     /// Size in base-asset 18-decimal fixed-point.
